% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-independencepvalue.Rmd.
\name{classical_p_val}
\alias{classical_p_val}
\title{Function to test the independence of two pre-specified groups of variables}
\usage{
classical_p_val(S, n, CP, k, mc_iter = 1000)
}
\arguments{
\item{S}{the covariance matrix of the data matrix \eqn{X}, where \eqn{X} = (\eqn{X_1} \eqn{X_2})}

\item{n}{the number of data points}

\item{CP}{the pre-specified grouping of the variables}

\item{k}{the group to be tested for independence with the remaining variables}

\item{mc_iter}{the number of Monte Carlo iterations to approximate the p-value}
}
\value{
The p-value for the test of independence.
}
\description{
This tests the null hypothesis of independence between two pre-specified groups of Gaussian variables. This function approximates the p-value corresponding to the Wilk's lambda statistics with Monte Carlo simulation with \code{mc_iter} iterations.
}
\examples{
# Simulates a 10 x 3 X_1 from N(0, I)
set.seed(1)
X_1 <- matrix(rnorm(30), 10, 3)

# Simulates a 10 x 2 X_2 from N(0, I) independently of X_1
set.seed(2)
X_2 <- matrix(rnorm(20), 10, 2)

# Compute the covariance matrix of X = (X_1 X_2).
covX <- cov(cbind(X_1, X_2))
# tests for a difference in means between X_1 and X_2
classical_p_val(S=covX, n=10, CP=c(rep(1, each = 3),rep(2, each = 2)), k=1, mc_iter=100)
}