% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCCA.R
\name{dCCA}
\alias{dCCA}
\title{Function to evaluate  the joint density of the canonical correlations}
\usage{
dCCA(n, p, rp, a, Lambda)
}
\arguments{
\item{n}{sample size}

\item{p}{total number of variables}

\item{rp}{minimum of the size of the two groups of variables.}

\item{a}{a scaling constant}

\item{Lambda}{the \code{rp} length vector where the joint density is to be evaluatedl \eqn{1 > Lambda[1]\ge Lambda[2] \ge ... \ge Lambda[rp] > 0}}
}
\value{
Scaled joined density of canonial correlations, evaluated at \eqn{Lambda}.
}
\description{
This function evaluates the joint density of the canonical correlations at a specific value \eqn{Lambda}.
}
\keyword{internal}
