% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_function_classical.R
\name{MC_function_classical}
\alias{MC_function_classical}
\title{Function for MC simulation for classical inference}
\usage{
MC_function_classical(p, rp, n)
}
\arguments{
\item{p}{p_1+p_2}

\item{rp}{min(p_1, p_2)}

\item{n}{sample size}
}
\value{
A sample from Wilks' lambda distribution
}
\description{
Given \eqn{p = p_1 + p2}, and \code{rp} = \eqn{min(p_1, p_2)}, and \code{n}, simulates from a Wilks' lambda distribution with appropriate parameters.
}
\keyword{internal}
