% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_function_selective.R
\name{MC_function_selective}
\alias{MC_function_selective}
\title{Function for MC simulation for selective inference}
\usage{
MC_function_selective(p, rp, n, A, b)
}
\arguments{
\item{p}{p_1+p_2}

\item{rp}{min(p_1, p_2)}

\item{n}{sample size}

\item{A}{matrix with \code{rp} rows}

\item{b}{numeric vector}
}
\value{
\item{statistic}{Test statistic corresponding to simulated \eqn{\Lambda}.}
\item{status}{Logical vector if the simulated \eqn{\Lambda} satisfies \eqn{A\Lambda \leq b}.}
}
\description{
For a given group of variables \eqn{P}, simulates from the joint distribution of the canonical correlations and check if \eqn{P} is recovered by applying \code{block_diag} on the perturbed covariance matrix. This condition can be characterized as the vector of canonical correlations \eqn{\Lambda} belonging in a convex polytope \eqn{A\Lambda \leq b}, where the matrix \eqn{A} and the vector \eqn{b} are functions of the threshold \eqn{c} and the sample covariance matrix.
}
\keyword{internal}
