% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-independencepvalue.Rmd.
\name{sample_canoncor_under_null}
\alias{sample_canoncor_under_null}
\title{Sample from the distribution of canonical correlations under the null}
\usage{
sample_canoncor_under_null(p, rp, n)
}
\arguments{
\item{p}{p_1+p_2}

\item{rp}{min(p_1, p_2)}

\item{n}{sample size}
}
\value{
A vector of length \code{rp} sampled from the joint distribution described
above.
}
\description{
Samples from the \code{rp}-dimensional joint distribution of the canonical
correlations between two groups of variables of size \code{p1} and \code{p2}, where
\code{rp} is the smaller of \code{p1} and \code{p2} and \code{p=p1+p2}, and \code{n} is the sample size.
}
\keyword{internal}