% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_psi.R
\name{sample_psi}
\alias{sample_psi}
\title{Sample from the distribution of eigenvalues of W * inv(T)}
\usage{
sample_psi(p, rp, n)
}
\arguments{
\item{p}{p_1+p_2}

\item{rp}{min(p_1, p_2)}

\item{n}{sample size}
}
\value{
A vector of length \code{rp} sampled from the joint distribution described
above.
}
\description{
Samples from the \code{rp}-dimensional joint distribution of the eigenvalues of
W * inv(T), where W and T are independent Wisharts with dimensions specified
in Prop 1(ii).  These are the {Psi_i}, and taking {sqrt(Psi_i/(1+Psi_i))}
gives a sample from the joint distirbution of the canonical correlations
between two groups of variables of size \code{p1} and \code{p2} under the null.
}
\keyword{internal}
