% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selective_p_val.R
\name{selective_p_val}
\alias{selective_p_val}
\title{Function to test the independence of a data-dependent group of variables with the remaining variables}
\usage{
selective_p_val(S, n, CP, c, k, d0 = 5, mc_iter = 1000)
}
\arguments{
\item{S}{the covariance matrix of the data matrix \eqn{X}, where \eqn{X} = (\eqn{X_1} \eqn{X_2})}

\item{n}{the number of data points}

\item{CP}{the grouping of the variables, an outcome of \code{block_diag}}

\item{c}{the threshold}

\item{k}{the group to be tested for independence with the remaining variables}

\item{d0}{a natural number; ff the number of canonical correlations is greater than \code{d0}, Monte Carlo simulation will be used to approximate the p-value for computational convenience; default value is 5}

\item{mc_iter}{the number of Monte Carlo iterations used to approximate the p-value}
}
\value{
The selective p-value for the test of independence.
}
\description{
This tests the null hypothesis of independence between two groups of Gaussian variables, where the groups are obtained via thresholding with \code{block_diag}.
}
\examples{
# Simulates a 10 x 5 X from N(0, I)
set.seed(1)
X <- matrix(rnorm(50), 10, 5)

# Compute the correlation matrix of X.
corX <- cor(X)
# Use 'block_diag' to obtain any block diagonal structure
block_diag_structure <- block_diag(corX, c= 0.5)
# test for independence of the variables in group 1 with the remaining variables
selective_p_val(S=cov(X), n=10, CP=block_diag_structure, c=0.5, k=1, d0=5, mc_iter=100)
}
