% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-independencepvalue.Rmd.
\name{selective_p_val_integrate}
\alias{selective_p_val_integrate}
\title{Compute selective p-value using numerical integration}
\usage{
selective_p_val_integrate(n, L, g, test_hyp)
}
\arguments{
\item{n}{sample size}

\item{L}{matrix used to define conditioning set}

\item{g}{vector used to define conditioning set}

\item{test_hyp}{output of \code{test_stat_CCA()}}
}
\description{
Compute selective p-value using numerical integration
}
\keyword{internal}